Add a boolean configuration option to disable reloading: grails.cache.ehcache.reloadable
Prefer the "ehcacheXmlLocation" over others so that if reloading is disabled, the ehcache configuration is still loaded

Ehcache doesn't support reloading of cache managers, which is what this plugin tries to do via private field modification and other tricks.
However, users of ehcache caches don't expect the caches to appear then be removed then be re-added.
To workaround potential problems with caches disappearing and re-appearing when reloading is enabled (such as Hibernate), proxy the returned Ehcache instance.